requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
Django==5.0.7
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
phonenumbers==8.13.42
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
whitenoise==6.7.0
zope.interface==6.4.post2
